* Once again, as usual, I always forget, don't commit while sleepy......
* Fixed the whole mqtthandler subscription and connection, forgot the goddam is_connected in every callback, promptly removed

* Implemented an intermediary configuration 'state', the server reads the config from user file as usual, and proceeds to read
if another exists in SVX_LOCAL_STATE_DIR with the name svxreflector.state, validates the file (with another Async::Config object,
and if exists, and is valid, overwrites the current running configuration.
This allows for changing configuration while running and persist changes through restarts. This feature is active through COMMAND_PTY
and MQTT, so that the server can be remote managed without restarts, to prevent connection losses and rebuild the connections on restarts.