Configuration state file saved/deleted based on configuration in memory.
When editing/adding configuration values through the PTY, the configuration in memory
changes, and there is no mechanism to retain those changes if a restart happend.
To remedy this, I implemented a state config in between the loading of the regular
user config file svxreflector.conf and another with all in memory configuration on
SVX_LOCAL_STATE_DIR/svxreflector.state which if present, will add/overwrite the user
config file.

To make things simple, the state file save the whole in memory configuration, thus,
it will overwrite all the user config if the same section/key exists in both,
maybe I'll eventually implement a way to figure out which configuration values are
more recent and always use the recent values.

This was implemented as a helper for the MQTT reflector patch, to aid on remote
management, and retain the latest configuration in check with remote applications
through mqtt when restarts happen, which can be avoided.